Merge pull request #1073 from RosettaCommons/aleaverfay/pose_builder_tests
Rosetta will now read ACE and NME residues without needing any special flags.

Changing the variant types given to the AcetylatedProteinNterm and MethylatedProteinCterm patches so that they register as LOWER_TERMINUS and UPPER_TERMINUS variants, which they are.

The effect is to enable some code in the PoseFromSFRBuilder that has been laying in wait.  Now, when Rosetta encounters an ACE residue on the Nterm of a chain, it will take the atoms for that residue and append them to the next residue, which will end up patched with the AcetylatedProteinNterm patch.

NOTE: we have two ACE patches in rosetta when we should really only have one.  "N_acetylated" and "AcetylatedProteinNterm". The more commonly used patch is "N_acetylated".  It, however, has a different set of backbone torsion angles than the one defined by "AcetylatedProteinNterm".  The same applies for the "C_methylamidated" and "MethylatedProteinCterm" pair.

We should rapidly decide how we are going to eliminate one of the two ACE patches and one of the two methylamidadtion patches, as duplication is the devil.

Old repository SHA1: 5f29cd737e2948d4669e707b4b5398901709616b