Merge pull request #1158 from RosettaCommons/xrw_PDB_diagnostics
(This is the PR opener, repeated).  This branch pulls in the PDB_diagnostic application written for the chemical XRW. The application just does the basic Rosetta exercises - score, repack, minimize. It also adds to the scorefile numbers of residues in all the categories supported directly by ResidueType. The key is that the application is not meant to run on a PDB (file), it's meant to run on THE PDB (all 115K+). There are two supports not in this PR - a data processing script in tools/ (will be merged there soon), and a godawful shell-script based job distributor that is robust against segfaults (will be merged noplace ever). Since we'll be discussing it at Winter RosettaCON I wanted it in master; also I'm tired of making special temp branches to merge this test code against the PDB-reading features it needs to test.

Note that the original history of the application was accidentally merged into an earlier XRW merge, but that merge was reverted after pushing. Thus this PR starts with an en-block reversion reversion to make the new code appear; its real history is further back.

The changes to not-obviously-this-application files are either cosmetic, or a bugfix in JD2 (boolean options should always check .value(), never .user(), so that the user can pass false...).

Integration and unit tests are a mess in master.  There are two integration test changes legitimately caused by this code: the new itest PDB_diagnostic (will fail because it's new), and central_class_modification (will fail because I changed JobDistributor [the bugfix noted above]).  

Old repository SHA1: 4417abefd0f2780b6e59c94418d9424614717511