Merge pull request #1206 from RosettaCommons/vmullig/phi_selector
Create a ResidueSelector for selecting alpha-amino acids in the positive- or negative-phi region of Ramachandran space



This is convenient for mixed D/L design: it's nice to be able to require that residues in the positive-phi region are only allowed to be D-amino acids, and that residues in the negative-phi region are only allowed to be L-amino acids. This can also be used for conventional design to ensure that positive-phi positions are only glycine, and negative-phi positions are never glycine. Or this can be used in conjunction with the aa_composition score term to limit the number of D-amino acids in the negative-phi region or L-amino acids in the positive-phi region. Or it can be used just to count residues in the positive-phi region as part of the process of filtering out bad designs.

So it should be useful.

@dabaker, @gbhardwaj, @hssnzdh2

Tasks:

    Add the ResidueSelector using Jared's nifty scripts.
    Actually write the apply() function.
    Write the parse_my_tag() function.
    ResidueSelectors require XSD stuff, don't they?
    Additional checks for termini and whatnot.
    Unit test.
    Documentation.
    Beauty.

Unit tests all pass on my machine.  No integration test or other test changes expected.  Merging.

Old repository SHA1: 0a819caf7e70949cd17e0a5889a1b5fe3c8e11b8