Merge pull request #1215 from RosettaCommons/aleaverfay/fix_matcher_chi_building_bug2
One-line fix to ProteinUpstreamBuilder::recover_hits to output NCAAs with discontiguous chis

This is the second half to the bug fix merged into master with PR #1085;
in that PR, ProteinUpstreamBuilder::build was updated to multiply the
pre_chitip_transform HomogeneousTransform when building the chitip
frame while actually building hits. There is a second place in the code,
however, where the geometries of the upstream residues are computed, and
this second location was not updated to use the new transforms.

This fixes the geometry of the NCAA in the test case given to me
by Sophia Tan. Thanks Sophia!

Old repository SHA1: ae4f6df348f840dcbd8e94c60751e644f0725194