Merge pull request #1236 from RosettaCommons/roccomoretti/silenttagfile
Add support for -in:file:tagfile to SilentFileInputter.

he relationship between -in:file:tags and -in:file:tagfile is similar to -in:file:s and -in:file:l -- that is, you can now pass a file with a list of tags, rather than having to list all the tags out on the commandline (and without the length limits doing so imposes).

The -in:file:tagfile is a pre-existing flag, only used once in AbrelaxApplication.cc, but with similar semantics. This extends the functionality to JD2 protocols. (Also fixes the option file description.)

Old repository SHA1: 4481504a3774db214ebda9eecefc40793e34c5aa