Merge pull request #1240 from RosettaCommons/vmullig/peppredict_setting
Fixing a minor bug in the simple_cycpep_predict app

When the user set certain residues to have certain dihedral values, the wrong residue was sometimes being set when the pose was circularly permuted. This affected the efficiency of sampling, but would not invalidate the scientific results (so this bugfix need not be pushed to BOINC immediately, @davidekim -- it can wait for whenever the next BOINC update happens).

Old repository SHA1: c7a4ed7af970318577287e38cd9f1105885e1e70