Merge pull request #1269 from RosettaCommons/jadolfbr/antibodies/skip_cis_trans_cluster_matching
Jadolfbr/antibodies/skip cis trans cluster matching

Adds an option to Rosetta Antibody for North/Dunbrack Clusters:

-allow_omega_mismatches_for_north_clusters
 - Skip first grouping by cis and trans designation for clusters in which a cis/trans designation currently does not exist.  

This is due to the North clusters being first grouped by the omega cis/trans for each residue, and then clusters within that (for example TTTTTTTCTT ).  This directly relates to the sequence profiles used for design.  We may eventually make this option default on, but for now it stays off.  If a cluster for a CDR cannot be found because of this mismatch, we give a nice warning to users in Red and let them know of this new option. 

Also add an integration test for identify_cdr_clusters. 

Thanks to @lyskov for helping with the tracer colors.  I added some info to the Tracer header that gives examples on how to do it for most common purposes.

Old repository SHA1: f7017b003254f119f4a24e536d979156b19363fa