Merge pull request #1326 from RosettaCommons/vmullig/write_template
Adding option to rosetta_scripts to print a template script

I often find myself going to the help wiki to find the template RosettaScript whenever I sit down to write a new script. If I'm working on a laptop without an Internet connection (e.g. on a train or a plane), this is a particular pain in the neck. This pull request adds a flag, -print_template_script, to the rosetta_scripts application that lets the application print out a blank, template RosettaScripts script and then exit. This should be handy for quick reference when sitting down to write a new script.

Tasks:
--    Add the feature.
--    Documentation.

Old repository SHA1: 7e3aa18d15737ae03a763739e8360a2ee33548eb