Merge pull request #1354 from RosettaCommons/smlewis/Extender_pilot_app_bitrot
trival bitrot repair on a senescent pilot app (CAP->COP); This is a noncompiling pilot app that was useful for generating starting structures for FloppyTail. As a noncompiling app it didn't get fixed when the smart pointers were updated and CAPs went away. This app remains disabled in scons (although it does compile now), so there oughtn't be test changes associated with it (and so I'm not loading the test server with it, either).

Old repository SHA1: 9a69209e4e20a15f3a46fd25c9fbef9f5169c305