Merge pull request #1410 from RosettaCommons/jadolfbr/doc_xrw2016
Overview
=======

This PR is part of the Documentation XRW 2016 effort and enables demo [integration] testing through ```integration.py```

This is done through an additional ```--demos``` and ```--tutorials``` cmd-line option, which indicate to the integration script to either run the demo or tutorial tests.  Additionally, single scripts can be passed at the end of the command line (Either names or relative paths or full paths or what have you.).

  
Tests are run for everything in the tutorials directory and public directory (which will change before the final merge.)

How it works
=========
If a ```command``` file is in the demo directory, we use that.  Alternatively, we parse all .md files looking for ```$>``` characters in the beginning of a (properly stripped) line, which indicates to test this command.  

If there are flags files, there can be additional flags.short files that the integration test is looking for in the command.  If these exist, it will add them to the command.  This is to enable short-testing runs, without having to maintain two separate versions of flags files. While some options cannot be overridden in this way, this works for most cases and is simple to maintain.  For example, this command in an md file (with a .short file in the directory):
```
$>../../../main/source/bin/score_jd2.default.linuxclangrelease -s 1ubq.pdb -out:suffix crystal @crystal_score_flags
```

Will be written to the output command file as:
```
%(bin)s/score_jd2.%(binext)s -s 1ubq.pdb -out:suffix crystal @crystal_score_flags @crystal_score_flags.short -database %(database)s -run:constant_seed -nodelay  2>&1 | egrep -vf path/to/ignore_list > log1
```

Individual tests can also be given, but the full relative path or the name of the demo will need to be given, for example,``` tutorials/Tutorial_4_Relax``` in addition to the ```--tutorials``` option (or ```--demos``` if this was a demo).

This should allow automated demo testing using our current integration test system, without forking integration.py.

Normal integration test changes
========================
Integration test changes are cosmetic and seem to be doing what that code block was originally intended to do.

Old repository SHA1: c6008f01ebca461840252b730ba0f6157dc6ed71

Co-authored-by: Vikram K. Mulligan <vmullig@u.washington.edu>