Merge pull request #1458 from RosettaCommons/vmullig/ncaa_betanov
Fix loading of NCAA rotamer libraries when using beta_nov15

The problem turned out to be that beta_nov15 turns on the Shapovalov rotamer libraries, which have a slightly different file format (an extra column in the middle of the file, which is irritating). When an NCAA library is loaded, it is therefore misinterpreted: Rosetta assumes that the extra column is there when it is not. The solution is to count the columns in the first line and to determine whether a file is a Shapovalov-format file or not.

Tasks:

-    Fix the bug.
-    Merge the latest master into branch.
-    Fix the fix so that it doesn't break the loading of gzipped rotamer libraries.
-    Beauty.

Old repository SHA1: 4c3be7cd05e33361cdbdd1f7bfbb666530440f85