Merge pull request #1472 from RosettaCommons/jadolfbr/jadolfbr/carbohydrates/val_fix
Fix simple glycosylation

Fix an uninitialized variable in the copy constructor of glycan relax that was causing the simple glycosylation unit test to fail in valgrind. Thanks to @roccomoretti for actually debugging this and finding where the variable was uninitialized in the first place.

Old repository SHA1: 892dc6f7fb355ad662847c82dc6bff29d39e2943