Merge pull request #1475 from RosettaCommons/vmullig/aspartimide
Add a score term to penalize sequences that are likely to result in aspartimide during peptide synthesis

In designing peptides, we're having to throw away or redesign a heck of a lot of designs because of DN, DS, DT, Dx, DG, dn, ds, dt, dX, or dG sequences, which often result in the unwanted aspartimide side-product during peptide synthesis. My Thursday night project will be to add a score term to penalize these two-residue sequences during design, so that the packer simply doesn't put them in. (There are several other things that we're also having to filter for, and each one cuts down our viable designs by an order of magnitude, greatly increasing the number of designs that we have to do in order to get one that we can synthesize. This is the easiest and quickest thing for me to fix with a non-filter-based solution, increasing our good design rate by an order of magnitude or so.)

Tasks:

-   Add the score term.
-   Have it only apply itself to pairs of residues that are conventionally connected (i.e. residue i is connected to residue j at residue i's C-terminus and residue j's N-terminus).
-   Have it penalize the sequences in question.
X  Double-check which sequences should be penalized. (Some also say DC, DD, and DR [and their mirrors]. DP and Dx', where x' is N-methylated, should be immune to aspartimide formation, though.) @gbhardwaj --> If there's a need to tweak what's penalized in the future, or to add general user controls, we can do this when we need it.
-    Unit test.
-    Beauty.
-    Documentation.

Old repository SHA1: 7ce9d5bf5174d497560ce1624761e0c166dc30c3