Merge pull request #1588 from RosettaCommons/jadolfbr/packstat
Fix packstat score test. 

So, the packstat score test was crazy with the JD2 merge (76 to >100k cycles or something).  Turns out that for packstat to SCORE, it needs to write out a PDB file to the output stream to turn it into a SimplePDB.  This caused the energies to fail when it tried to access the weights DURING scoring (which are now output by default). 

 * Use StructFileRepOptions to turn off output of the energies table (as well as any pose cache). This should fix the test. 
 * Use StructFileRepOptions to remove global setting of OptionSystem data Yay! @aleaverfay 


Turns out that packstat is only really tested in the score performance test.  If you care about packstat, you should add a proper test for it (and probably refactor it to not output an entire PDB from a pose at each score cycle).  

@smlewis @roccomoretti @willsheffler 

Thanks for the help in sleuthing Steven!

Old repository SHA1: 6dace0b774a15197b55060860756a0087e20f00a