Merge pull request #1581 from RosettaCommons/jadolfbr/grafting_disulfides
Fix antibody.cc

This PR adds an option, optimal_graft, that fixes the grafting component of antibody.cc.  By fix, I mean that it uses the exact settings we have successfully used for antibody design - this results in 100% graft closure, clean disulfides, and lower energies.  By using other code for this option, we also fix a major bug in the grafting component of antibody.cc, which is responsible for much of the disulfide issues.  

We also add disulfide score terms for centroid mode in AnchoredGraftMover's default, smooth centroid score function.  They help to lessen the effects of the bug without the optimal_graft option, but do not alleviate it completely.  

This will change a few integration tests that use AnchoredGraftMover or CCDEndsGraftMover including antibody_designer and AnchoredPDBCreator due to the additional centroid score terms.

Thanks to @lqtza for testing and debugging all this with me, as well as the work calling Relax directly for the final relax, instead of generating it from cmd-line options.

Old repository SHA1: 4075d4caf7624cc1bf6f04f9913d8ee9a2e948da

Co-authored-by: Jeliazko Jeliazkov <jeliazkov@jhu.edu>