Merge pull request #1596 from RosettaCommons/roccomoretti/relax_centroid_native
Fix issue with centroid poses and -relax:constrain_relax_to_native_coords

-relax:constrain_relax_to_native_coords would only work if you also included the -in:file:fullatom flag. Otherwise, the native pose would get loaded as centroid mode. (By set_native_in_mover() in src/protocols/jd2/util.cc)

That's terribly silly, as relax is a full-atom only protocol. (Doubly so, as the -s gets loaded as fullatom by default, but the -in:file:native is loaded as centroid.)

This pull request bypasses the "is it centroid or not" logic, and always loads the native reference pose as fullatom.

Integration test changes expected in relax_w_allatom_cst, as the native pose is now loaded as centroid (due to removing the -in:file:fullatom flag from the integration test),  even though the centroid version is not used.

Old repository SHA1: 0ca4248d244ad1929a9b195d6730421130a12652