Merge pull request #1503 from RosettaCommons/roccomoretti/rts_type
Separate the concept of a ResidueTypeSet "type" from a RTS "name"

There is (or should be) a distinction between the "type" of a ResidueTypeSet (e.g. "Fullatom"/"Centoid") and the "name" of the ResidueTypeSet. For example, we have an fa_standard RT, but doesn't that imply there could be an fa_nonstandard?

This pull request attempts to separate out the concept of a TypeSet type (as an enum), and replace the instances of using names as comparisons with types (where it makes sense). This means that you can now ask "is this residue a 'fullatom' residue type, rather than asking "is this ResidueType a member of the fa_standard ResidueTypeSet.

A bunch of integration test changes expected, as I made SwitchResidueTypeSet much less chatty when it's "switching" a pose that already has residues of the desired type.

Old repository SHA1: af7a7aa6e1dc2a7d5d256ef96f939ea6b153ae88