Merge pull request #1680 from RosettaCommons/vmullig/brian_sequence_bug
Fixing inverted PDBInfo check in Sequence.cc

Brian Koepnick and Jeff Flatten of the Foldit team found a bug in Sequence.cc: the existence of a PDBInfo object is checked before the object is accessed, but the check is inverted. That is, the object is accessed ONLY if it doesn't exist, which makes no sense. This fixes that. Thanks to these two for finding this.

@bkoepnick @jflat06

Old repository SHA1: 29b1d1ae2cc811e97b8ff17a834a688df4af4cd1