Merge pull request #1712 from RosettaCommons/vmullig/revert_cart_breakages
Reverting pull request 1699 due to breakages of simple_cycpep_predict  tests.

@BrandonFrenz: Fix these test breakages in your branch, test on the test server, and then re-merge with master.

And the community as a whole needs to start paying more attention to the test server. We had green lights across the board for some time, but now we're starting to ignore breakages. It is NOT a developer's responsibility to keep re-fixing his or her features that someone else has broken, IF those features are covered with tests. It is the rest of the community's responsibility to make sure that they don't break tests.

Old repository SHA1: e747805349759f67c22fa1ae72750950215a82b5