Merge pull request #1735 from RosettaCommons/JWLabonte/PyRosetta
PyRosetta: Restoring pose_from_pdb()

Its absence is wrecking a whole bunch of user's scripts and it is what is published in the Workshops. We can't just go changing the names of functions in PyRosetta so rapidly, so I am restoring the wrapper I added in version 3 that was commented out in version 4.

Old repository SHA1: 0c3dc27b1bc25aebf9ac631db8150d6719e41eee