Merge pull request #1221 from RosettaCommons/roccomoretti/cst_output
Add Constraint info application and mover

So you've generated a bunch of models, and you want to see how well they match the constraints. Certainly you can get the complete constraint score, but what if you want to know how each constraint behaves?

This pull request adds a mover and an application which will take one or more constraint files and one or more input structures, and will produces an output table of each constraint and subconstraint (for nested constraints like Ambigious and KofN constraints) regarding a) what their score is, and b) what is the metric prior to function application. The RosettaScripts interface can also be set to report on the constraints that are currently in the pose. 

There's a slight change to the Constraint object interface to enable (and enforce) the ability to do this. 

Old repository SHA1: 344661fce99d4d823d3fe86facc82e221f8a0db8