Merge pull request #1086 from RosettaCommons/JWLabonte/enzymatic_movers
EnzymaticMovers: GlycosyltransferaseMover example

This merge introduces a framework for a new concept in Rosetta for `EnzymaticMover`s, `Mover`s that simulate enzymatic reactions of post-translational modifications to `Pose`s, which I introduced at Winter RosettaCON2016.

(If interested, view my presentation, which should be available from the Rosetta Commons wiki, or contact me. In short, and `EnzymaticMover` moves "`ResidueType` space" in a `Pose` instead of torsion space and creates starting structures for modeling rather than ending decoys from modeling.)

This is a work-in-progress.

In its current stage, it only uses sequence information, but it will be expanded to include structural information for the determination of which sites on a `Pose` will be modified.

Currently, I have only added a generic, human, glycosyltransferase as an example, which modifies NXS/T sites.

This merge includes unit and integration tests.

It also, as an aside, includes a sorting of the init files for `Mover`s as a bonus. (Please keep those files clean!)

All tests pass.

Old repository SHA1: d2d8e4f18962824632aa714132e36f5503caaa6d