Merge pull request #1858 from RosettaCommons/everyday847/farfar_metrics
There may be a curious issue with the way division is working with the fractional scores -- but we'll only know for sure after we use it a lot, and we won't be able to until we've got these features to play with. So, merging for now!

Old repository SHA1: e459c5251c69a54603143858726b763fe0617170